pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel", "versioningit"]
build-backend = "setuptools.build_meta"

[project]
name = "dotbins"
description = "Keep updated binaries in your dotfiles"
dynamic = ["version"]
authors = [{ name = "Bas Nijholt", email = "bas@nijho.lt" }]
dependencies = [
    "requests",
    "pyyaml",
    "rich>=13.9.4",
]
requires-python = ">=3.9"

[project.readme]
file = "README.md"
content-type = "text/markdown"

[project.urls]
Homepage = "https://github.com/basnijholt/dotbins"

[project.optional-dependencies]
test = ["pytest", "pre-commit", "coverage", "pytest-cov"]
docs = []

[project.scripts]
dotbins = "dotbins.cli:main"

[tool.setuptools.packages.find]
include = ["unidep.*", "unidep"]

[tool.setuptools.package-data]
"unidep" = ["py.typed"]

[tool.pytest.ini_options]
addopts = """
    --cov=dotbins
    --cov-report term
    --cov-report html
    --cov-fail-under=70
"""

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]

[tool.black]
line_length = 88

[tool.ruff]
line-length = 150
target-version = "py39"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "T20",     # flake8-print
    "ANN101",  # Missing type annotation for {name} in method
    "S101",    # Use of assert detected
    "PD901",   # df is a bad variable name. Be kinder to your future self.
    "ANN401",  # Dynamically typed expressions (typing.Any) are disallowed in {name}
    "D402",    # First line should not be the function's signature
    "PLW0603", # Using the global statement to update `X` is discouraged
    "D401",    # First line of docstring should be in imperative mood
    "SLF001",  # Private member accessed
    "PTH",     # Use pathlib.Path
    "DTZ005",  # The use of `datetime.datetime.now()` without `tz` argument is not allowed
    "PLR0913", # Too many arguments to function call
    "S602",    # `subprocess` call with `shell=True` identified, security issue
    "S603",    # `subprocess` call: check for execution of untrusted input
    "S202",    # Uses of `tarfile.extractall()`
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["SLF001"]
".github/*" = ["INP001"]

[tool.ruff.lint.mccabe]
max-complexity = 18

[tool.mypy]
python_version = "3.9"

[tool.versioningit]

[tool.versioningit.vcs]
method = "git"
match = ["v*"]
default-tag = "0.0.0"

[dependency-groups]
dev = [
    "pytest-cov>=6.0.0",
    "pytest>=8.3.5",
    "requests-mock>=1.12.1",
    "pytest-mock>=3.14.0",
    "ruff>=0.9.10",
]
